replace forloop solc fixture type with test-case macro (#9841)
For all tests of revm instructions replaced `for fixture_type` with
test-case macro

---------

+++ substrate/frame/revive/fixtures/contracts/BlockInfo.sol
@@ -2,31 +2,31 @@
 pragma solidity ^0.8.0;
 
 contract BlockInfo {
-    function blockNumber() public view returns (uint) {
-        return block.number;
+    function blockNumber() public view returns (uint64) {
+        return uint64(block.number);
     }
 
     function coinbase() public view returns (address) {
         return block.coinbase;
     }
 
-    function timestamp() public view returns (uint) {
-        return block.timestamp;
+    function timestamp() public view returns (uint64) {
+        return uint64(block.timestamp);
     }
 
-    function difficulty() public view returns (uint) {
-        return block.difficulty;
+    function difficulty() public view returns (uint64) {
+        return uint64(block.difficulty);
     }
 
-    function gaslimit() public view returns (uint) {
-        return block.gaslimit;
+    function gaslimit() public view returns (uint64) {
+        return uint64(block.gaslimit);
     }
 
-    function chainid() public view returns (uint) {
-        return block.chainid;
+    function chainid() public view returns (uint64) {
+        return uint64(block.chainid);
     }
 
-    function basefee() public view returns (uint) {
-        return block.basefee;
+    function basefee() public view returns (uint64) {
+        return uint64(block.basefee);
     }
 }
